Iterate through list of eids; removed ClearEntity.
We're iterating through a list of eids (via domain_view) rather than
domain_view itself, because eventually we'll want to delete components
inside of a process -- and that will be super-hard if we're changing the
iteration size every time we do it. Keeping the setup and teardown
methods for processes just in case, though.